add start at correct _start address

+++ fib.c
@@ -1,7 +1,7 @@
-int _start() {
-    int a = 0, b = 1, c = 0;
+long _start() {
+    long a = 0, b = 1, c = 0;
 
-    for (int i = 0; i < 10; i++) {
+    for (long i = 0; i < 10; i++) {
         c = a + b;
         a = b;
         b = c;
